+++ src/UnitSpacing.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bakame\Stackwatch;
+
+enum UnitSpacing
+{
+    case None;
+    case Space;
+}